+++ libs/cardano-ledger-api/CHANGELOG.md
@@ -1,8 +1,12 @@
 # Version history for `cardano-ledger-api`
 
-## 1.13.0.1
+## 1.13.1.0
 
-*
+* Add `querySetSnapshotStakePoolDistr` (`GetPoolDistr2`).
+* Add `queryDRepDelegatees` (`GetFilteredVoteDelegatees`).
+* Add `queryStakePoolRelays` (`GetLedgerPeerSnapshot`).
+* Add `queryAccountsDeposits` (`GetStakeDelegDeposits`).
+* Add `queryCurrentEpochNo` (`GetEpochNo`).
 
 ## 1.13.0.0